created new seperate test file for testing Configuration class, added an add class to concatinate configuration objects with build in extened stringifications for substitusions as a default behavior
