+++ tests/test_loaders.py
@@ -35,58 +35,3 @@ def test_loader_dot_notation():
      actual = props.jobname
 
      assert actual == expected
-
-
-def test_configuration_no_schema():
-    config = Configuration(os.path.join(solo_file_base, 'config.yaml'))
-    assert len(config)
-
-
-def test_configuration_with_schema_ok():
-    config = Configuration(os.path.join(solo_file_base, 'config.yaml'), 'config_schema')
-    assert len(config)
-
-def test_YAML_loader_parse_env():
-
-    props = Configuration(pathlib.Path(os.path.join(uwtools_file_base,"fixtures/experiment.yaml")),environment=True)
-
-    expected = os.environ.get('USER')
-    actual = props.user
-    assert actual == expected
-
-
-def test_YAML_loader_update():
-
-    props = Configuration(pathlib.Path(os.path.join(uwtools_file_base,"fixtures/experiment.yaml")))
-    props2 = Configuration(pathlib.Path(os.path.join(uwtools_file_base,"fixtures/gfs.yaml")))
-
-    config = Template.substitute_structure( props2, TemplateConstants.DOLLAR_PARENTHESES, props.get)
-
-    expected =  "/home/myexpid/{{current_cycle}}"
-    actual = config.datapath
-
-    assert actual == expected
-
-def test_YAML_loader_inplace_update():
-
-    props = Configuration(pathlib.Path(os.path.join(uwtools_file_base,"fixtures/gfs.yaml")))
-    config = Template.substitute_structure( props, TemplateConstants.DOLLAR_PARENTHESES, props.get)
-
-    expected =  "testpassed"
-    actual = props.testupdate
-
-    assert actual == expected
-
-def test_YAML_loader_realtime_update():
-
-    props = Configuration(pathlib.Path(os.path.join(uwtools_file_base,"fixtures/experiment.yaml")))
-    props2 = Configuration(pathlib.Path(os.path.join(uwtools_file_base,"fixtures/gfs.yaml")))
-    config = Template.substitute_structure( props2, TemplateConstants.DOLLAR_PARENTHESES, props.get)
-    config = Template.substitute_structure( config, TemplateConstants.DOUBLE_CURLY_BRACES, props.get)
-
-    print(config)
-
-    expected =  "/home/myexpid/10102022"
-    actual = config.updated_datapath
-
-    assert actual == expected
